Move `package` tests to `execution`
